RFC: split mpl_interface into two modules to enable delaying matplotlib import even when it's available
