+++ setup.cfg
@@ -19,7 +19,7 @@ universal = 1
 test = pytest
 
 [tool:pytest]
-norecursedirs = benchmarks paper .*
+norecursedirs = benchmarks paper .* unyt/_mpl_array_converter
 
 [versioneer]
 VCS = git
